Refactor item search control methods into findable module
